メール認証関連をAPIにも追加

+++ app/Http/Controllers/Api/CampScheduleController.php
@@ -3,10 +3,10 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Api\CampScheduleRequest;
 use App\Http\Requests\Api\CampScheduleStoreRequest;
 use App\Http\Requests\Api\CampScheduleUpdateRequest;
 use App\Repositories\CampSchedule\CampScheduleRepositoryInterface;
+use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Http\Request;
 
 class CampScheduleController extends Controller
@@ -31,7 +31,11 @@ class CampScheduleController extends Controller
      */
     public function getList(Request $request)
     {
-        $campPlaces = $this->campScheduleRepository->getListByUserId(
+        if ($request->user() instanceof MustVerifyEmail && !$request->user()->hasVerifiedEmail()) {
+            return response()->json(['code' => 403, 'error_message' => 'メール認証が完了していません'], 403, [], JSON_PRETTY_PRINT);
+        }
+
+        $campSchedules = $this->campScheduleRepository->getListByUserId(
             $request->user()->id,
             $request->offset ?? 0,
             $request->limit ?? 10,
@@ -40,7 +44,7 @@ class CampScheduleController extends Controller
 
         return response()->json([
             'code' => 200,
-            'campPlaces' => $campPlaces,
+            'campSchedules' => $campSchedules,
         ], 200, [], JSON_PRETTY_PRINT);
     }
 
@@ -64,6 +68,10 @@ class CampScheduleController extends Controller
      */
     public function store(CampScheduleStoreRequest $request)
     {
+        if ($request->user() instanceof MustVerifyEmail && !$request->user()->hasVerifiedEmail()) {
+            return response()->json(['code' => 403, 'error_message' => 'メール認証が完了していません'], 403, [], JSON_PRETTY_PRINT);
+        }
+
         try {
             $campSchedule = $this->campScheduleRepository->store($request->all() + ['user_id' => $request->user()->id]);
 
@@ -86,6 +94,10 @@ class CampScheduleController extends Controller
      */
     public function update(int $id, CampScheduleUpdateRequest $request)
     {
+        if ($request->user() instanceof MustVerifyEmail && !$request->user()->hasVerifiedEmail()) {
+            return response()->json(['code' => 403, 'error_message' => 'メール認証が完了していません'], 403, [], JSON_PRETTY_PRINT);
+        }
+
         try {
             $campSchedule = $this->campScheduleRepository->update($id, $request->except(['user_id']));
 
@@ -103,10 +115,15 @@ class CampScheduleController extends Controller
 
     /**
      * @param int $id
+     * @param Request $request
      * @return \Illuminate\Http\JsonResponse
      */
-    public function destroy(int $id)
+    public function destroy(int $id, Request $request)
     {
+        if ($request->user() instanceof MustVerifyEmail && !$request->user()->hasVerifiedEmail()) {
+            return response()->json(['code' => 403, 'error_message' => 'メール認証が完了していません'], 403, [], JSON_PRETTY_PRINT);
+        }
+
         try {
             $this->campScheduleRepository->destroy($id);
